pyproject.toml
[tool.poetry]
name = "fastapi-graphql-blog"
version = "0.1.0"
description = "FastAPI + Graphql blog api"
authors = ["ullyzian <aviev2000@icloud.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
graphene = "^2.1.8"
uvicorn = "^0.13.3"
sqlalchemy = "^1.3.22"
graphene-sqlalchemy = "^2.3.0"
python-dotenv = "^0.15.0"
alembic = "^1.5.1"
passlib = "^1.7.4"
psycopg2-binary = "^2.8.6"
psycopg2 = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
